requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Detection ─────────────────────────────────────────────────────────────────
ultralytics==8.2.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python-headless>=4.9.0
Pillow>=10.2.0

# ── Data / Viz ────────────────────────────────────────────────────────────────
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── App ───────────────────────────────────────────────────────────────────────
streamlit>=1.33.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
PyYAML>=6.0.1
